requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pandas>=2.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
langchain-community>=0.0.28
langchain-google-genai>=0.0.13
duckduckgo-search>=3.9.6
python-dotenv>=1.0.0
aiofiles>=23.2.1
python-multipart>=0.0.9
httpx>=0.27.0
python-dotenv
gunicorn
